Support statically linking CUDA runtime for Java bindings (#9873)
Fixes #9528.  Adds the ability to statically link the CUDA runtime when building the Java bindings.  This also adds the ability for the Java bindings to link against an archive of libcudf rather than the shared library.  This is recommended when statically linking the CUDA runtime, as it reduces the number of shared libraries that pull in the CUDA runtime.

The Java CI script to build the official artifact has been updated to building libcudf as an archive and statically linking the CUDA runtime by default.

Authors:
  - Jason Lowe (https://github.com/jlowe)

Approvers:
  - Robert (Bobby) Evans (https://github.com/revans2)

URL: https://github.com/rapidsai/cudf/pull/9873