Add zlib to cudfjni link when using static libcudf library dependency (#9890)
When the Java bindings use libcudf as a static library dependency, the resulting link is missing zlib as a dynamic link dependency.  This can result in runtime errors such as `undefined symbol: inflateInit2_` when trying to load a file using gzip compression.  This adds zlib to the cudfjni link when using libcudf.a as the cudf dependency.

Authors:
  - Jason Lowe (https://github.com/jlowe)

Approvers:
  - Alessandro Bellina (https://github.com/abellina)
  - Nghia Truong (https://github.com/ttnghia)
  - Gera Shegalov (https://github.com/gerashegalov)

URL: https://github.com/rapidsai/cudf/pull/9890