JNI: Function to copy and set validity from bool column. (#9901)
This commit introduces a JNI function `ColumnVector#copyWithValidity(ColumnView boolean_column)` to use a `BOOL8` column as the validity buffer of a column.

As per the convention in the JNI layer, this method does not modify the current object. A new `ColumnVector` is returned, containing the contents of the current object. The validity of the returned `ColumnVector` is set to the `DeviceBuffer` representation of the logical contents of the supplied `BOOL8` column.

This method will be useful to set a column's validity based on a column resulting from a boolean operation.

Authors:
  - MithunR (https://github.com/mythrocks)

Approvers:
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/9901