Use upstream clang-format pre-commit hook. (#10659)
This PR improves the use of clang-format in pre-commit and CI by using a hook that is managed and cached by pre-commit. Developers who do not use pre-commit will see no changes to their workflow. Developers using pre-commit will be guaranteed that their code is formatted with the pinned version of clang-format (and will always align with CI's style check expectations) even if they have a different version of clang-format (or no version of clang-format) installed on their path. Like all pre-commit hooks for linting CMake and Python code, this version pinning will need to be kept in sync with the conda environment if we choose to upgrade clang-format.

Note: this is aligned with how cuCollections uses clang-format, and it seems that has gone well. Thanks to @PointKernel for testing it on that repo!
https://github.com/NVIDIA/cuCollections/blob/dev/.pre-commit-config.yaml

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Yunsong Wang (https://github.com/PointKernel)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/10659