Fix construction of nested structs with EMPTY child (#10761)
This PR adds the ability to construct a structs columns with a parent null mask, but an EMPTY child. Prior to this PR, this would fail with a runtime error. This is because the parent null mask is superimposed on the EMPTY child, but EMPTY columns cannot contain null masks.

This PR also includes a small change in the Cython, where we map libcudf's `EMPTY` type to a `int8` type on the Python side.

Authors:
  - Ashwin Srinath (https://github.com/shwina)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - David Wendt (https://github.com/davidwendt)
  - Michael Wang (https://github.com/isVoid)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/10761