Fix index alignment for Series objects with repeated index (#11103)
The logic for aligning the indexes of two Series objects with a repeated value of the index is broken. We previously introduced an optimization that avoids a join operation, but unfortunately, that doesn't work in every situation, as seen in https://github.com/rapidsai/cudf/issues/11094. This PR removes that special-casing.

Closes #11094

Authors:
  - Ashwin Srinath (https://github.com/shwina)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/11103