Optimization to gpu::PreprocessColumnData in parquet reader. (#11252)
When handling list columns in the parquet reader, we have to run a preprocess step that computes several things per-page before we can decode values.  If the user has further specified artificial row bounds (`skip_rows`, `min_rows`) we have to do a second pass during the preprocess step.

If the user has _not_ specified row bounds, there is no need to do this; however the code was naively always doing so.  This PR simply detects when we're reading all rows (which is 99% of use cases) and skips the second pass.

Also includes some cleanup of redundant stream synchronizations.

Also worth mentioning, this `skip_rows`/`num_rows` feature is going to be deprecated in 22.08 so we will be able to follow up further in 22.10 to rip more of this code out.

Authors:
  - https://github.com/nvdbaranec

Approvers:
  - Mike Wilson (https://github.com/hyperbolic2346)
  - Jim Brennan (https://github.com/jbrennan333)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/11252